backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
exceptiongroup==1.2.2
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2024.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gradio_client==1.4.2
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper @ git+https://github.com/openai/whisper.git@173ff7dd1d9fb1c4fddea0d41d704cfefeb8908c
packaging==24.2
pillow==10.2.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pytube==15.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.10.2
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.0
triton==3.1.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.3
websockets==12.0
Werkzeug==3.1.3
youtube-transcript-api==0.6.2
yt-dlp==2024.11.4
